legacy branch: merge r2662-r2759
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/branches/legacy@2778 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/OpenCapDeviceDlg.h
@@ -45,7 +45,7 @@ public:
 	CComboBox m_countryctrl;
 
 	CString m_vidstr, m_audstr;
-	long m_country;
+	int m_country;
 
 // Dialog Data
 	enum { IDD = IDD_OPENCAPDEVICE_DLG };
